+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 -----------------------
 
 - Add form error if source file contains invalid header (#1780)
+- Fix for incorrect header order on 'confirm' page (#1786)
 - Remove unneeded format method overrides (#1785)
 - Support dynamic selection of Resource class based on request property ()